+++ examples/basic-app/src/bootstrap/app.ts
@@ -25,7 +25,7 @@ export default class {
 
         h3App.use((event) => kernel.handle(event, async () => undefined))
 
-        serve(h3App, { port: 3000 })
+        serve(h3App, { port: 3001 })
         console.log('🚀 H3ravel running at http://localhost:3000')
     }
 }