+++ examples/basic-app/tsup.config.ts
@@ -2,9 +2,9 @@ import { defineConfig } from 'tsup';
 
 const env = process.env.NODE_ENV || 'development'
 const postCmd = env === 'development'
-    ? `&& NODE_ENV=${env} SRC_PATH=dist node -r tsconfig-paths/register dist/server.js`
-    : ''
-
+    ? `&& NODE_ENV=${env} node -r tsconfig-paths/register src/server.ts`
+    : `&& NODE_ENV=${env} SRC_PATH=dist node -r tsconfig-paths/register dist/server.js`
+console.log(env, '==============')
 export default defineConfig({
     entry: ['src/**/*.ts'],
     format: ['esm'],