+++ packages/core/src/Providers/CoreServiceProvider.ts
@@ -1,5 +1,6 @@
 import 'reflect-metadata'
 
+import { Registerer } from '../Registerer';
 import { ServiceProvider } from '../ServiceProvider'
 
 /**
@@ -15,6 +16,6 @@ export class CoreServiceProvider extends ServiceProvider {
     public static priority = 999;
 
     register () {
-        // Core bindings
+        Registerer.register()
     }
 }