+++ packages/core/src/Registerer.ts
@@ -1,0 +1,8 @@
+import { dd, dump } from '@h3ravel/support'
+
+export class Registerer {
+    static register () {
+        globalThis.dd = dd
+        globalThis.dump = dump
+    }
+}