+++ packages/support/src/Helpers/DumpDie.ts
@@ -9,6 +9,11 @@ const inspect = (thing: any) => {
     })
 }
 
+/**
+ * Dump something and kill the process for quick debugging. Based on Laravel's dd()
+ * 
+ * @param args 
+ */
 export const dd = (...args: unknown[]): never => {
     args.forEach((thing) => {
         console.log(inspect(thing))
@@ -17,6 +22,11 @@ export const dd = (...args: unknown[]): never => {
     process.exit(1)
 }
 
+/**
+ * Dump something but keep the process for quick debugging. Based on Laravel's dump()
+ * 
+ * @param args 
+ */
 export const dump = (...args: unknown[]): void => {
     args.forEach((thing) => {
         console.log(inspect(thing))